move public and tweak build
